requirements.txt
notion-client==2.2.1
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.2.0
python-dotenv==1.0.0
requests==2.31.0
fastapi
uvicorn[standard]
sqlalchemy
authlib
itsdangerous
httpx
